Zoho
Inventory management is a systematic approach to sourcing, storing, and selling inventory—both raw materials (components) and finished goods (products). In business terms, inventory management means the right stock, at the right levels, in the right place, at the right time, and at the right cost as well as price.